[delete]使っていないクラスを削除

+++ Src/Utility/StringUtility.cpp
@@ -9,7 +9,7 @@ std::wstring StringUtility::StringToWstring(const std::string& str)
     auto result = MultiByteToWideChar(CP_ACP,
         MB_PRECOMPOSED | MB_ERR_INVALID_CHARS,
         str.c_str(),
-        str.length(),
+        static_cast<int>(str.length()),
         nullptr,
         0);
 
@@ -20,9 +20,9 @@ std::wstring StringUtility::StringToWstring(const std::string& str)
     result = MultiByteToWideChar(CP_ACP,
         MB_PRECOMPOSED | MB_ERR_INVALID_CHARS,
         str.c_str(),//“ü—Í•¶Žš—ñ
-        str.length(),
+        static_cast<int>(str.length()),
         ret.data(),
-        ret.size());
+        static_cast<int>(ret.size()));
 
     return ret;
 }
@@ -35,7 +35,7 @@ std::string StringUtility::WstringToString(const std::wstring& wstr)
         CP_ACP,
         0,
         wstr.c_str(),//“ü—Í•¶Žš—ñ
-        wstr.length(),
+        static_cast<int>(wstr.length()),
         nullptr,
         0,
         nullptr,
@@ -47,9 +47,9 @@ std::string StringUtility::WstringToString(const std::wstring& wstr)
         CP_ACP,
         0,
         wstr.c_str(),//“ü—Í•¶Žš—ñ
-        wstr.length(),
+        static_cast<int>(wstr.length()),
         ret.data(),
-        ret.size(),
+        static_cast<int>(ret.size()),
         nullptr,
         nullptr);
     return ret;
